Move new putMetrics method into CloudWatch trait
