+++ common/app/metrics/DurationMetric.scala
@@ -1,10 +1,8 @@
 package metrics
 
 import akka.agent.Agent
-import com.amazonaws.services.cloudwatch.model.{Dimension, StandardUnit, MetricDatum, PutMetricDataRequest}
+import com.amazonaws.services.cloudwatch.model.StandardUnit
 import common.AkkaAgent
-import model.diagnostics.CloudWatch
-import scala.collection.JavaConversions._
 import org.joda.time.DateTime
 
 trait DataPoint {
@@ -30,30 +28,6 @@ case class DurationMetric(name: String, metricUnit: StandardUnit) extends Fronte
 
   def recordDuration(timeInMillis: Long) = record(DurationDataPoint(timeInMillis, Option(DateTime.now)))
 
-}
-
-object CloudWatchMetrics {
-  def putMetrics(metrics: List[FrontendMetric], dimensions: List[Dimension]): Unit = {
-    for {
-      metric <- metrics
-      dataPoints <- metric.getDataPoints.grouped(20)
-      dataPoint <- dataPoints
-    } {
-      val request = new PutMetricDataRequest()
-        .withNamespace("Application")
-        .withMetricData {
-          val metricDatum = new MetricDatum()
-            .withValue(dataPoint.value)
-            .withUnit(metric.metricUnit)
-            .withMetricName(metric.name)
-            .withDimensions(dimensions)
-
-          dataPoint.time.fold(metricDatum) { t => metricDatum.withTimestamp(t.toDate)}
-        }
-
-      CloudWatch.cloudwatch.putMetricDataAsync(request)
-    }
-  }
 }
 
 object UkPressLatencyMetric extends DurationMetric("uk-press-latency", StandardUnit.Milliseconds)
