+++ common/app/model/diagnostics/CloudWatch.scala
@@ -4,6 +4,7 @@ import com.amazonaws.services.cloudwatch.AmazonCloudWatchAsyncClient
 import com.amazonaws.handlers.AsyncHandler
 import conf.Configuration
 import com.amazonaws.services.cloudwatch.model._
+import metrics.FrontendMetric
 import scala.collection.JavaConversions._
 import common.Logging
 import Configuration._
@@ -50,6 +51,29 @@ trait CloudWatch extends Logging {
 
   def putWithDimensions(namespace: String, metrics: Map[String, Double], dimensions: Seq[Dimension]) =
     put(namespace, metrics, Seq(stage) ++ dimensions)
+
+
+  def putMetrics(metrics: List[FrontendMetric], dimensions: List[Dimension]): Unit = {
+    for {
+      metric <- metrics
+      dataPoints <- metric.getDataPoints.grouped(20)
+      dataPoint <- dataPoints
+    } {
+      val request = new PutMetricDataRequest()
+        .withNamespace("Application")
+        .withMetricData {
+        val metricDatum = new MetricDatum()
+          .withValue(dataPoint.value)
+          .withUnit(metric.metricUnit)
+          .withMetricName(metric.name)
+          .withDimensions(dimensions)
+
+        dataPoint.time.fold(metricDatum) { t => metricDatum.withTimestamp(t.toDate)}
+      }
+
+      CloudWatch.cloudwatch.putMetricDataAsync(request)
+    }
+  }
 
 }
